Let's go

+++ src/motor/Draw.hpp
@@ -5,5 +5,5 @@
 // Implementations are in Rendering.cpp
 namespace Draw
 {
-    void Sprite(f32 x, f32 y);
+    void Sprite(int32 x, int32 y);
 }
